+++ core/src/main/java/com/github/shadowsocks/bg/BaseService.kt
@@ -39,9 +39,11 @@ import com.github.shadowsocks.plugin.PluginManager
 import com.github.shadowsocks.utils.Action
 import com.github.shadowsocks.utils.broadcastReceiver
 import com.github.shadowsocks.utils.printLog
+import com.github.shadowsocks.utils.readableMessage
 import com.google.firebase.analytics.FirebaseAnalytics
 import kotlinx.coroutines.*
 import java.io.File
+import java.net.BindException
 import java.net.InetAddress
 import java.net.URL
 import java.net.UnknownHostException
@@ -51,20 +53,22 @@ import java.util.*
  * This object uses WeakMap to simulate the effects of multi-inheritance.
  */
 object BaseService {
-    /**
-     * IDLE state is only used by UI and will never be returned by BaseService.
-     */
-    const val IDLE = 0
-    const val CONNECTING = 1
-    const val CONNECTED = 2
-    const val STOPPING = 3
-    const val STOPPED = 4
+    enum class State(val canStop: Boolean = false) {
+        /**
+         * Idle state is only used by UI and will never be returned by BaseService.
+         */
+        Idle,
+        Connecting(true),
+        Connected(true),
+        Stopping,
+        Stopped,
+    }
 
     const val CONFIG_FILE = "shadowsocks.conf"
     const val CONFIG_FILE_UDP = "shadowsocks-udp.conf"
 
     class Data internal constructor(private val service: Interface) {
-        var state = STOPPED
+        var state = State.Stopped
         var processes: GuardedProcessPool? = null
         var proxy: ProxyInstance? = null
         var udpFallback: ProxyInstance? = null
@@ -81,7 +85,7 @@ object BaseService {
         val binder = Binder(this)
         var connectingJob: Job? = null
 
-        fun changeState(s: Int, msg: String? = null) {
+        fun changeState(s: State, msg: String? = null) {
             if (state == s && msg == null) return
             binder.stateChanged(s, msg)
             state = s
@@ -102,7 +106,7 @@ object BaseService {
         private val bandwidthListeners = mutableMapOf<IBinder, Long>()  // the binder is the real identifier
         private val handler = Handler()
 
-        override fun getState(): Int = data!!.state
+        override fun getState(): Int = data!!.state.ordinal
         override fun getProfileName(): String = data!!.proxy?.profile?.name ?: "Idle"
 
         override fun registerCallback(cb: IShadowsocksServiceCallback) {
@@ -124,12 +128,12 @@ object BaseService {
             handler.postDelayed(this::onTimeout, bandwidthListeners.values.min() ?: return)
         }
         private fun onTimeout() {
-            val proxies = listOfNotNull(data!!.proxy, data!!.udpFallback)
+            val proxies = listOfNotNull(data?.proxy, data?.udpFallback)
             val stats = proxies
                     .map { Pair(it.profile.id, it.trafficMonitor?.requestUpdate()) }
                     .filter { it.second != null }
                     .map { Triple(it.first, it.second!!.first, it.second!!.second) }
-            if (stats.any { it.third } && state == CONNECTED && bandwidthListeners.isNotEmpty()) {
+            if (stats.any { it.third } && data!!.state == State.Connected && bandwidthListeners.isNotEmpty()) {
                 val sum = stats.fold(TrafficStats()) { a, b -> a + b.second }
                 broadcast { item ->
                     if (bandwidthListeners.contains(item.asBinder())) {
@@ -145,7 +149,7 @@ object BaseService {
             val wasEmpty = bandwidthListeners.isEmpty()
             if (bandwidthListeners.put(cb.asBinder(), timeout) == null) {
                 if (wasEmpty) registerTimeout()
-                if (state != CONNECTED) return
+                if (data!!.state != State.Connected) return
                 var sum = TrafficStats()
                 val proxy = data!!.proxy ?: return
                 proxy.trafficMonitor?.out.also { stats ->
@@ -177,9 +181,9 @@ object BaseService {
             callbacks.unregister(cb)
         }
 
-        fun stateChanged(s: Int, msg: String?) {
+        fun stateChanged(s: State, msg: String?) {
             val profileName = profileName
-            broadcast { it.stateChanged(s, profileName, msg) }
+            broadcast { it.stateChanged(s.ordinal, profileName, msg) }
         }
 
         fun trafficPersisted(ids: List<Long>) {
@@ -211,9 +215,9 @@ object BaseService {
                 return
             }
             val s = data.state
-            when (s) {
-                STOPPED -> startRunner()
-                CONNECTED -> stopRunner(true)
+            when {
+                s == State.Stopped -> startRunner()
+                s.canStop -> stopRunner(true)
                 else -> Crashlytics.log(Log.WARN, tag, "Illegal state when invoking use: $s")
             }
         }
@@ -249,11 +253,12 @@ object BaseService {
         }
 
         fun stopRunner(restart: Boolean = false, msg: String? = null) {
-            if (data.state == STOPPING) return
+            if (data.state == State.Stopping) return
             // channge the state
-            data.changeState(STOPPING)
+            data.changeState(State.Stopping)
             GlobalScope.launch(Dispatchers.Main, CoroutineStart.UNDISPATCHED) {
                 Core.analytics.logEvent("stop", bundleOf(Pair(FirebaseAnalytics.Param.METHOD, tag)))
+                data.connectingJob?.cancelAndJoin() // ensure stop connecting first
                 this@Interface as Service
                 // we use a coroutineScope here to allow clean-up in parallel
                 coroutineScope {
@@ -278,7 +283,7 @@ object BaseService {
                 }
 
                 // change the state
-                data.changeState(STOPPED, msg)
+                data.changeState(State.Stopped, msg)
 
                 // stop the service if nothing has bound to it
                 if (restart) startRunner() else stopSelf()
@@ -291,7 +296,7 @@ object BaseService {
 
         fun onStartCommand(intent: Intent?, flags: Int, startId: Int): Int {
             val data = data
-            if (data.state != STOPPED) return Service.START_NOT_STICKY
+            if (data.state != State.Stopped) return Service.START_NOT_STICKY
             val profilePair = Core.currentProfile
             this as Context
             if (profilePair == null) {
@@ -318,7 +323,7 @@ object BaseService {
             data.notification = createNotification(profile.formattedName)
             Core.analytics.logEvent("start", bundleOf(Pair(FirebaseAnalytics.Param.METHOD, tag)))
 
-            data.changeState(CONNECTING)
+            data.changeState(State.Connecting)
             data.connectingJob = GlobalScope.launch(Dispatchers.Main) {
                 try {
                     Executable.killAll()    // clean up old processes
@@ -328,8 +333,7 @@ object BaseService {
 
                     data.processes = GuardedProcessPool {
                         printLog(it)
-                        data.connectingJob?.cancelAndJoin()
-                        stopRunner(false, it.localizedMessage)
+                        stopRunner(false, it.readableMessage)
                     }
                     startProcesses()
 
@@ -337,16 +341,18 @@ object BaseService {
                     data.udpFallback?.scheduleUpdate()
                     RemoteConfig.fetch()
 
-                    data.changeState(CONNECTED)
+                    data.changeState(State.Connected)
                 } catch (_: CancellationException) {
                     // if the job was cancelled, it is canceller's responsibility to call stopRunner
                 } catch (_: UnknownHostException) {
                     stopRunner(false, getString(R.string.invalid_server))
                 } catch (exc: Throwable) {
-                    if (exc !is PluginManager.PluginNotFoundException && exc !is VpnService.NullConnectionException) {
+                    if (exc !is PluginManager.PluginNotFoundException &&
+                            exc !is BindException &&
+                            exc !is VpnService.NullConnectionException) {
                         printLog(exc)
                     }
-                    stopRunner(false, "${getString(R.string.service_failed)}: ${exc.localizedMessage}")
+                    stopRunner(false, "${getString(R.string.service_failed)}: ${exc.readableMessage}")
                 } finally {
                     data.connectingJob = null
                 }