+++ mobile/src/main/java/com/github/shadowsocks/MainActivity.kt
@@ -23,6 +23,7 @@ package com.github.shadowsocks
 import android.app.Activity
 import android.app.backup.BackupManager
 import android.content.ActivityNotFoundException
+import android.content.DialogInterface
 import android.content.Intent
 import android.net.VpnService
 import android.nfc.NdefMessage
@@ -30,6 +31,7 @@ import android.nfc.NfcAdapter
 import android.os.Bundle
 import android.os.DeadObjectException
 import android.os.Handler
+import android.os.Parcelable
 import android.util.Log
 import android.view.KeyCharacterMap
 import android.view.KeyEvent
@@ -51,6 +53,8 @@ import com.github.shadowsocks.aidl.TrafficStats
 import com.github.shadowsocks.bg.BaseService
 import com.github.shadowsocks.database.Profile
 import com.github.shadowsocks.database.ProfileManager
+import com.github.shadowsocks.plugin.AlertDialogFragment
+import com.github.shadowsocks.plugin.Empty
 import com.github.shadowsocks.preference.DataStore
 import com.github.shadowsocks.preference.OnPreferenceDataStoreChangeListener
 import com.github.shadowsocks.utils.Key
@@ -58,6 +62,7 @@ import com.github.shadowsocks.widget.ServiceButton
 import com.github.shadowsocks.widget.StatsBar
 import com.google.android.material.navigation.NavigationView
 import com.google.android.material.snackbar.Snackbar
+import kotlinx.android.parcel.Parcelize
 
 class MainActivity : AppCompatActivity(), ShadowsocksConnection.Callback, OnPreferenceDataStoreChangeListener,
         NavigationView.OnNavigationItemSelectedListener {
@@ -65,7 +70,18 @@ class MainActivity : AppCompatActivity(), ShadowsocksConnection.Callback, OnPref
         private const val TAG = "ShadowsocksMainActivity"
         private const val REQUEST_CONNECT = 1
 
-        var stateListener: ((Int) -> Unit)? = null
+        var stateListener: ((BaseService.State) -> Unit)? = null
+    }
+
+    @Parcelize
+    data class ProfilesArg(val profiles: List<Profile>) : Parcelable
+    class ImportProfilesDialogFragment : AlertDialogFragment<ProfilesArg, Empty>() {
+        override fun AlertDialog.Builder.prepare(listener: DialogInterface.OnClickListener) {
+            setTitle(R.string.add_profile_dialog)
+            setPositiveButton(R.string.yes) { _, _ -> arg.profiles.forEach { ProfileManager.createProfile(it) } }
+            setNegativeButton(R.string.no, null)
+            setMessage(arg.profiles.joinToString("\n"))
+        }
     }
 
     // UI
@@ -91,12 +107,12 @@ class MainActivity : AppCompatActivity(), ShadowsocksConnection.Callback, OnPref
     }
 
     // service
-    var state = BaseService.IDLE
-    override fun stateChanged(state: Int, profileName: String?, msg: String?) = changeState(state, msg, true)
+    var state = BaseService.State.Idle
+    override fun stateChanged(state: BaseService.State, profileName: String?, msg: String?) = changeState(state, msg, true)
     override fun trafficUpdated(profileId: Long, stats: TrafficStats) {
         if (profileId == 0L) this@MainActivity.stats.updateTraffic(
                 stats.txRate, stats.rxRate, stats.txTotal, stats.rxTotal)
-        if (state != BaseService.STOPPING) {
+        if (state != BaseService.State.Stopping) {
             (supportFragmentManager.findFragmentById(R.id.fragment_holder) as? ToolbarFragment)
                     ?.onTrafficUpdated(profileId, stats)
         }
@@ -105,7 +121,7 @@ class MainActivity : AppCompatActivity(), ShadowsocksConnection.Callback, OnPref
         ProfilesFragment.instance?.onTrafficPersisted(profileId)
     }
 
-    private fun changeState(state: Int, msg: String? = null, animate: Boolean = false) {
+    private fun changeState(state: BaseService.State, msg: String? = null, animate: Boolean = false) {
         fab.changeState(state, animate)
         stats.changeState(state)
         if (msg != null) snackbar(getString(R.string.vpn_error, msg)).show()
@@ -115,7 +131,7 @@ class MainActivity : AppCompatActivity(), ShadowsocksConnection.Callback, OnPref
     }
 
     private fun toggle() = when {
-        state == BaseService.CONNECTED -> Core.stopService()
+        state.canStop -> Core.stopService()
         DataStore.serviceMode == Key.modeVpn -> {
             val intent = VpnService.prepare(this)
             if (intent != null) startActivityForResult(intent, REQUEST_CONNECT)
@@ -127,11 +143,11 @@ class MainActivity : AppCompatActivity(), ShadowsocksConnection.Callback, OnPref
     private val handler = Handler()
     private val connection = ShadowsocksConnection(handler, true)
     override fun onServiceConnected(service: IShadowsocksService) = changeState(try {
-        service.state
+        BaseService.State.values()[service.state]
     } catch (_: DeadObjectException) {
-        BaseService.IDLE
+        BaseService.State.Idle
     })
-    override fun onServiceDisconnected() = changeState(BaseService.IDLE)
+    override fun onServiceDisconnected() = changeState(BaseService.State.Idle)
     override fun onBinderDied() {
         connection.disconnect(this)
         connection.connect(this, this)
@@ -152,7 +168,7 @@ class MainActivity : AppCompatActivity(), ShadowsocksConnection.Callback, OnPref
         super.onCreate(savedInstanceState)
         setContentView(R.layout.layout_main)
         stats = findViewById(R.id.stats)
-        stats.setOnClickListener { if (state == BaseService.CONNECTED) stats.testConnection() }
+        stats.setOnClickListener { if (state == BaseService.State.Connected) stats.testConnection() }
         drawer = findViewById(R.id.drawer)
         navigation = findViewById(R.id.navigation)
         navigation.setNavigationItemSelectedListener(this)
@@ -164,7 +180,7 @@ class MainActivity : AppCompatActivity(), ShadowsocksConnection.Callback, OnPref
         fab = findViewById(R.id.fab)
         fab.setOnClickListener { toggle() }
 
-        changeState(BaseService.IDLE)   // reset everything to init state
+        changeState(BaseService.State.Idle) // reset everything to init state
         connection.connect(this, this)
         DataStore.publicStore.registerChangeListener(this)
 
@@ -188,17 +204,8 @@ class MainActivity : AppCompatActivity(), ShadowsocksConnection.Callback, OnPref
         }
         if (sharedStr.isNullOrEmpty()) return
         val profiles = Profile.findAllUrls(sharedStr, Core.currentProfile?.first).toList()
-        if (profiles.isEmpty()) {
-            snackbar().setText(R.string.profile_invalid_input).show()
-            return
-        }
-        AlertDialog.Builder(this)
-                .setTitle(R.string.add_profile_dialog)
-                .setPositiveButton(R.string.yes) { _, _ -> profiles.forEach { ProfileManager.createProfile(it) } }
-                .setNegativeButton(R.string.no, null)
-                .setMessage(profiles.joinToString("\n"))
-                .create()
-                .show()
+        if (profiles.isEmpty()) snackbar().setText(R.string.profile_invalid_input).show()
+        else ImportProfilesDialogFragment().withArg(ProfilesArg(profiles)).show(supportFragmentManager, null)
     }
 
     override fun onPreferenceDataStoreChanged(store: PreferenceDataStore, key: String?) {