+++ mobile/src/main/java/com/github/shadowsocks/ProfileConfigFragment.kt
@@ -22,23 +22,24 @@ package com.github.shadowsocks
 
 import android.app.Activity
 import android.content.BroadcastReceiver
+import android.content.DialogInterface
 import android.content.Intent
 import android.os.Bundle
+import android.os.Parcelable
 import android.view.MenuItem
 import androidx.appcompat.app.AlertDialog
 import androidx.preference.*
 import com.github.shadowsocks.Core.app
 import com.github.shadowsocks.database.Profile
 import com.github.shadowsocks.database.ProfileManager
-import com.github.shadowsocks.plugin.PluginConfiguration
-import com.github.shadowsocks.plugin.PluginContract
-import com.github.shadowsocks.plugin.PluginManager
-import com.github.shadowsocks.plugin.PluginOptions
+import com.github.shadowsocks.plugin.*
 import com.github.shadowsocks.preference.*
 import com.github.shadowsocks.utils.Action
 import com.github.shadowsocks.utils.DirectBoot
 import com.github.shadowsocks.utils.Key
+import com.github.shadowsocks.utils.readableMessage
 import com.google.android.material.snackbar.Snackbar
+import kotlinx.android.parcel.Parcelize
 
 class ProfileConfigFragment : PreferenceFragmentCompat(),
         Preference.OnPreferenceChangeListener, OnPreferenceDataStoreChangeListener {
@@ -46,6 +47,20 @@ class ProfileConfigFragment : PreferenceFragmentCompat(),
         override fun provideSummary(preference: EditTextPreference?) = "\u2022".repeat(preference?.text?.length ?: 0)
 
         private const val REQUEST_CODE_PLUGIN_CONFIGURE = 1
+        const val REQUEST_UNSAVED_CHANGES = 2
+    }
+
+    @Parcelize
+    data class ProfileIdArg(val profileId: Long) : Parcelable
+    class DeleteConfirmationDialogFragment : AlertDialogFragment<ProfileIdArg, Empty>() {
+        override fun AlertDialog.Builder.prepare(listener: DialogInterface.OnClickListener) {
+            setTitle(R.string.delete_confirm_prompt)
+            setPositiveButton(R.string.yes) { _, _ ->
+                ProfileManager.delProfile(arg.profileId)
+                requireActivity().finish()
+            }
+            setNegativeButton(R.string.no, null)
+        }
     }
 
     private var profileId = -1L
@@ -114,7 +129,7 @@ class ProfileConfigFragment : PreferenceFragmentCompat(),
         }.show(fragmentManager ?: return, Key.pluginConfigure)
     }
 
-    fun saveAndExit() {
+    private fun saveAndExit() {
         val profile = ProfileManager.getProfile(profileId) ?: Profile()
         profile.id = profileId
         profile.deserialize()
@@ -140,7 +155,7 @@ class ProfileConfigFragment : PreferenceFragmentCompat(),
         DataStore.dirty = true
         true
     } catch (exc: RuntimeException) {
-        Snackbar.make(view!!, exc.localizedMessage, Snackbar.LENGTH_LONG).show()
+        Snackbar.make(view!!, exc.readableMessage, Snackbar.LENGTH_LONG).show()
         false
     }
 
@@ -167,28 +182,26 @@ class ProfileConfigFragment : PreferenceFragmentCompat(),
     }
 
     override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
-        if (requestCode == REQUEST_CODE_PLUGIN_CONFIGURE) when (resultCode) {
-            Activity.RESULT_OK -> {
-                val options = data?.getStringExtra(PluginContract.EXTRA_OPTIONS)
-                pluginConfigure.text = options
-                onPreferenceChange(null, options)
+        when (requestCode) {
+            REQUEST_CODE_PLUGIN_CONFIGURE -> when (resultCode) {
+                Activity.RESULT_OK -> {
+                    val options = data?.getStringExtra(PluginContract.EXTRA_OPTIONS)
+                    pluginConfigure.text = options
+                    onPreferenceChange(null, options)
+                }
+                PluginContract.RESULT_FALLBACK -> showPluginEditor()
             }
-            PluginContract.RESULT_FALLBACK -> showPluginEditor()
-        } else super.onActivityResult(requestCode, resultCode, data)
+            REQUEST_UNSAVED_CHANGES -> when (resultCode) {
+                DialogInterface.BUTTON_POSITIVE -> saveAndExit()
+                DialogInterface.BUTTON_NEGATIVE -> requireActivity().finish()
+            }
+            else -> super.onActivityResult(requestCode, resultCode, data)
+        }
     }
 
     override fun onOptionsItemSelected(item: MenuItem) = when (item.itemId) {
         R.id.action_delete -> {
-            val activity = requireActivity()
-            AlertDialog.Builder(activity)
-                    .setTitle(R.string.delete_confirm_prompt)
-                    .setPositiveButton(R.string.yes) { _, _ ->
-                        ProfileManager.delProfile(profileId)
-                        activity.finish()
-                    }
-                    .setNegativeButton(R.string.no, null)
-                    .create()
-                    .show()
+            DeleteConfirmationDialogFragment().withArg(ProfileIdArg(profileId)).show(this)
             true
         }
         R.id.action_apply -> {