+++ mobile/src/main/java/com/github/shadowsocks/QuickToggleShortcut.kt
@@ -55,14 +55,15 @@ class QuickToggleShortcut : Activity(), ShadowsocksConnection.Callback {
     }
 
     override fun onServiceConnected(service: IShadowsocksService) {
-        when (service.state) {
-            BaseService.STOPPED -> Core.startService()
-            BaseService.CONNECTED -> Core.stopService()
+        val state = BaseService.State.values()[service.state]
+        when {
+            state.canStop -> Core.stopService()
+            state == BaseService.State.Stopped -> Core.startService()
         }
         finish()
     }
 
-    override fun stateChanged(state: Int, profileName: String?, msg: String?) { }
+    override fun stateChanged(state: BaseService.State, profileName: String?, msg: String?) { }
 
     override fun onDestroy() {
         connection.disconnect(this)