+++ tv/src/main/java/com/github/shadowsocks/tv/MainPreferenceFragment.kt
@@ -30,7 +30,6 @@ import android.os.DeadObjectException
 import android.os.Handler
 import android.text.format.Formatter
 import android.util.Log
-import android.util.LongSparseArray
 import android.widget.Toast
 import androidx.leanback.preference.LeanbackPreferenceFragmentCompat
 import androidx.lifecycle.Observer
@@ -44,7 +43,6 @@ import com.github.shadowsocks.aidl.IShadowsocksService
 import com.github.shadowsocks.aidl.ShadowsocksConnection
 import com.github.shadowsocks.aidl.TrafficStats
 import com.github.shadowsocks.bg.BaseService
-import com.github.shadowsocks.database.Profile
 import com.github.shadowsocks.database.ProfileManager
 import com.github.shadowsocks.net.HttpsTest
 import com.github.shadowsocks.net.TcpFastOpen
@@ -54,13 +52,13 @@ import com.github.shadowsocks.preference.PortPreferenceListener
 import com.github.shadowsocks.utils.Key
 import com.github.shadowsocks.utils.datas
 import com.github.shadowsocks.utils.printLog
-import org.json.JSONArray
+import com.github.shadowsocks.utils.readableMessage
 
 class MainPreferenceFragment : LeanbackPreferenceFragmentCompat(), ShadowsocksConnection.Callback,
         OnPreferenceDataStoreChangeListener {
     companion object {
         private const val REQUEST_CONNECT = 1
-        private const val REQUEST_IMPORT_PROFILES = 2
+        private const val REQUEST_REPLACE_PROFILES = 2
         private const val REQUEST_EXPORT_PROFILES = 3
         private const val TAG = "MainPreferenceFragment"
     }
@@ -88,9 +86,9 @@ class MainPreferenceFragment : LeanbackPreferenceFragmentCompat(), ShadowsocksCo
     private lateinit var tester: HttpsTest
 
     // service
-    var state = BaseService.IDLE
+    var state = BaseService.State.Idle
         private set
-    override fun stateChanged(state: Int, profileName: String?, msg: String?) = changeState(state, msg)
+    override fun stateChanged(state: BaseService.State, profileName: String?, msg: String?) = changeState(state, msg)
     override fun trafficUpdated(profileId: Long, stats: TrafficStats) {
         if (profileId == 0L) requireContext().let { context ->
             this.stats.summary = getString(R.string.stat_summary,
@@ -101,26 +99,26 @@ class MainPreferenceFragment : LeanbackPreferenceFragmentCompat(), ShadowsocksCo
         }
     }
 
-    private fun changeState(state: Int, msg: String? = null) {
-        fab.isEnabled = state == BaseService.STOPPED || state == BaseService.CONNECTED
+    private fun changeState(state: BaseService.State, msg: String? = null) {
+        fab.isEnabled = state.canStop || state == BaseService.State.Stopped
         fab.setTitle(when (state) {
-            BaseService.CONNECTING -> R.string.connecting
-            BaseService.CONNECTED -> R.string.stop
-            BaseService.STOPPING -> R.string.stopping
+            BaseService.State.Connecting -> R.string.connecting
+            BaseService.State.Connected -> R.string.stop
+            BaseService.State.Stopping -> R.string.stopping
             else -> R.string.connect
         })
         stats.setTitle(R.string.connection_test_pending)
-        stats.isVisible = state == BaseService.CONNECTED
-        if (state != BaseService.CONNECTED) {
+        stats.isVisible = state == BaseService.State.Connected
+        if (state != BaseService.State.Connected) {
             trafficUpdated(0, TrafficStats())
             tester.status.removeObservers(this)
-            if (state != BaseService.IDLE) tester.invalidate()
+            if (state != BaseService.State.Idle) tester.invalidate()
         } else tester.status.observe(this, Observer {
             it.retrieve(stats::setTitle) { Toast.makeText(requireContext(), it, Toast.LENGTH_LONG).show() }
         })
         if (msg != null) Toast.makeText(requireContext(), getString(R.string.vpn_error, msg), Toast.LENGTH_SHORT).show()
         this.state = state
-        if (state == BaseService.STOPPED) {
+        if (state == BaseService.State.Stopped) {
             controlImport.isEnabled = true
             tfo.isEnabled = true
             serviceMode.isEnabled = true
@@ -141,11 +139,11 @@ class MainPreferenceFragment : LeanbackPreferenceFragmentCompat(), ShadowsocksCo
     private val handler = Handler()
     private val connection = ShadowsocksConnection(handler, true)
     override fun onServiceConnected(service: IShadowsocksService) = changeState(try {
-        service.state
+        BaseService.State.values()[service.state]
     } catch (_: DeadObjectException) {
-        BaseService.IDLE
+        BaseService.State.Idle
     })
-    override fun onServiceDisconnected() = changeState(BaseService.IDLE)
+    override fun onServiceDisconnected() = changeState(BaseService.State.Idle)
     override fun onBinderDied() {
         connection.disconnect(requireContext())
         connection.connect(requireContext(), this)
@@ -203,7 +201,7 @@ class MainPreferenceFragment : LeanbackPreferenceFragmentCompat(), ShadowsocksCo
         }
 
         tester = ViewModelProviders.of(this).get()
-        changeState(BaseService.IDLE)   // reset everything to init state
+        changeState(BaseService.State.Idle) // reset everything to init state
         connection.connect(requireContext(), this)
         DataStore.publicStore.registerChangeListener(this)
     }
@@ -228,7 +226,7 @@ class MainPreferenceFragment : LeanbackPreferenceFragmentCompat(), ShadowsocksCo
 
     fun startService() {
         when {
-            state != BaseService.STOPPED -> return
+            state != BaseService.State.Stopped -> return
             DataStore.serviceMode == Key.modeVpn -> {
                 val intent = VpnService.prepare(requireContext())
                 if (intent != null) startActivityForResult(intent, REQUEST_CONNECT)
@@ -254,7 +252,7 @@ class MainPreferenceFragment : LeanbackPreferenceFragmentCompat(), ShadowsocksCo
 
     override fun onPreferenceTreeClick(preference: Preference?) = when (preference?.key) {
         Key.id -> {
-            if (state == BaseService.CONNECTED) Core.stopService()
+            if (state == BaseService.State.Connected) Core.stopService()
             true
         }
         Key.controlStats -> {
@@ -263,15 +261,14 @@ class MainPreferenceFragment : LeanbackPreferenceFragmentCompat(), ShadowsocksCo
         }
         Key.controlImport -> {
             startFilesForResult(Intent(Intent.ACTION_GET_CONTENT).apply {
-                addCategory(Intent.CATEGORY_OPENABLE)
-                type = "application/json"
+                type = "application/*"
                 putExtra(Intent.EXTRA_ALLOW_MULTIPLE, true)
-            }, REQUEST_IMPORT_PROFILES)
+                putExtra(Intent.EXTRA_MIME_TYPES, arrayOf("application/*", "text/*"))
+            }, REQUEST_REPLACE_PROFILES)
             true
         }
         Key.controlExport -> {
             startFilesForResult(Intent(Intent.ACTION_CREATE_DOCUMENT).apply {
-                addCategory(Intent.CATEGORY_OPENABLE)
                 type = "application/json"
                 putExtra(Intent.EXTRA_TITLE, "profiles.json")   // optional title that can be edited
             }, REQUEST_EXPORT_PROFILES)
@@ -280,9 +277,9 @@ class MainPreferenceFragment : LeanbackPreferenceFragmentCompat(), ShadowsocksCo
         else -> super.onPreferenceTreeClick(preference)
     }
 
-    private fun startFilesForResult(intent: Intent?, requestCode: Int) {
+    private fun startFilesForResult(intent: Intent, requestCode: Int) {
         try {
-            startActivityForResult(intent, requestCode)
+            startActivityForResult(intent.addCategory(Intent.CATEGORY_OPENABLE), requestCode)
             return
         } catch (_: ActivityNotFoundException) { } catch (_: SecurityException) { }
         Toast.makeText(requireContext(), R.string.file_manager_missing, Toast.LENGTH_SHORT).show()
@@ -294,41 +291,30 @@ class MainPreferenceFragment : LeanbackPreferenceFragmentCompat(), ShadowsocksCo
                 Toast.makeText(requireContext(), R.string.vpn_permission_denied, Toast.LENGTH_SHORT).show()
                 Crashlytics.log(Log.ERROR, TAG, "Failed to start VpnService from onActivityResult: $data")
             }
-            REQUEST_IMPORT_PROFILES -> {
+            REQUEST_REPLACE_PROFILES -> {
                 if (resultCode != Activity.RESULT_OK) return
-                val profiles = ProfileManager.getAllProfiles()?.associateBy { it.formattedAddress }
-                val feature = profiles?.values?.singleOrNull { it.id == DataStore.profileId }
-                val lazyClear = lazy { ProfileManager.clear() }
                 val context = requireContext()
-                for (uri in data!!.datas) try {
-                    Profile.parseJson(context.contentResolver.openInputStream(uri)!!.bufferedReader().readText(),
-                            feature) {
-                        lazyClear.value
-                        // if two profiles has the same address, treat them as the same profile and copy stats over
-                        profiles?.get(it.formattedAddress)?.apply {
-                            it.tx = tx
-                            it.rx = rx
-                        }
-                        ProfileManager.createProfile(it)
-                    }
+                try {
+                    ProfileManager.createProfilesFromJson(data!!.datas.asSequence().map {
+                        context.contentResolver.openInputStream(it)
+                    }, true)
                 } catch (e: Exception) {
                     printLog(e)
-                    Toast.makeText(context, e.localizedMessage, Toast.LENGTH_SHORT).show()
+                    Toast.makeText(context, e.readableMessage, Toast.LENGTH_SHORT).show()
                 }
                 populateProfiles()
             }
             REQUEST_EXPORT_PROFILES -> {
                 if (resultCode != Activity.RESULT_OK) return
-                val profiles = ProfileManager.getAllProfiles()
+                val profiles = ProfileManager.serializeToJson()
                 val context = requireContext()
                 if (profiles != null) try {
-                    val lookup = LongSparseArray<Profile>(profiles.size).apply { profiles.forEach { put(it.id, it) } }
                     context.contentResolver.openOutputStream(data?.data!!)!!.bufferedWriter().use {
-                        it.write(JSONArray(profiles.map { it.toJson(lookup) }.toTypedArray()).toString(2))
+                        it.write(profiles.toString(2))
                     }
                 } catch (e: Exception) {
                     printLog(e)
-                    Toast.makeText(context, e.localizedMessage, Toast.LENGTH_SHORT).show()
+                    Toast.makeText(context, e.readableMessage, Toast.LENGTH_SHORT).show()
                 }
             }
             else -> super.onActivityResult(requestCode, resultCode, data)